Moved proof generator to common
